# public/api-docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Cipi API",
...
        "version": "1.20.0",
        "contact": {
            "name": "Cipi",
            "url": "https://cipi.sh"
        },
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/apps/{name}/redirects": {
            "get": {
                "tags": [
                    "Redirects"
                ],
                "summary": "List app + path redirects",
                "description": "Whole-app redirect and path redirects of an app (`cipi redirect list --json`). Synchronous. Requires `redirects-view` ability and Cipi CLI >= 5.4.1.",
                "operationId": "redirectList",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RedirectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "503": {
                        "description": "Cipi CLI error (message in `error`)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Redirects"
                ],
                "summary": "Add / update a path redirect",
                "description": "Add (or update) a path redirect (`cipi redirect add`). A `from` ending in `/` is a prefix match: the rest of the request URI is appended to the target (`--no-path` behaviour via `keep_path: false`). `to` is a same-app `/path` or an `http(s)://` URL. Reserved paths (ACME, `/cipi/webhook`, Reverb, `/index.php`, ...) and colliding rules are refused by the CLI. Synchronous with nginx test + revert. Requires `redirects-manage` ability and Cipi CLI >= 5.4.1.",
                "operationId": "redirectAdd",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RedirectAddRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated redirects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RedirectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Cipi CLI error (message in `error`)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Redirects"
                ],
                "summary": "Remove a path redirect",
                "description": "Remove a path redirect by its `from` path (`cipi redirect remove`). Requires `redirects-manage` ability and Cipi CLI >= 5.4.1.",
                "operationId": "redirectRemove",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RedirectRemoveRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated redirects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RedirectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Cipi CLI error (message in `error`)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/apps/{name}/redirect": {
            "put": {
                "tags": [
                    "Redirects"
                ],
                "summary": "Set the whole-app redirect",
                "description": "Every hostname of the app (www host included) redirects to the target URL in one hop (`cipi redirect set`). Path and query are kept by default (`keep_path`). ACME stays public so certificates keep renewing; a target served by the app itself is refused (loop). Path redirects and proxies keep working (more specific than `location /`). Requires `redirects-manage` ability and Cipi CLI >= 5.4.1.",
                "operationId": "redirectSet",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RedirectSetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated redirects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RedirectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Cipi CLI error (message in `error`)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Redirects"
                ],
                "summary": "Unset the whole-app redirect",
                "description": "Remove the whole-app redirect and forget its target (`cipi redirect unset`). Requires `redirects-manage` ability and Cipi CLI >= 5.4.1.",
                "operationId": "redirectUnset",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated redirects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RedirectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "503": {
                        "description": "Cipi CLI error (message in `error`)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum API token. Create with: cipi api token create"
            }
        },
        "parameters": {
            "AppName": {
                "name": "name",
                "in": "path",
                "required": true,
                "description": "App username (e.g. myapp)",
                "schema": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9]{2,31}$"
                }
            },
...
        },
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                }
            },
...
            "AppRedirect": {
                "type": "object",
                "nullable": true,
                "description": "Whole-app redirect (null when none is saved)",
                "properties": {
                    "enabled": {
                        "type": "boolean",
                        "example": true
                    },
                    "to": {
                        "type": "string",
                        "example": "https://new.example.com"
                    },
                    "code": {
                        "type": "integer",
                        "enum": [
                            301,
                            302,
                            307,
                            308
                        ],
                        "example": 301
                    },
                    "keep_path": {
                        "type": "boolean",
                        "description": "Whether the request path and query are appended to the target",
                        "example": true
                    }
                }
            },
            "RedirectRule": {
                "type": "object",
                "properties": {
                    "from": {
                        "type": "string",
                        "description": "Source path. Ending in / = prefix match",
                        "example": "/blog/"
                    },
                    "to": {
                        "type": "string",
                        "description": "Target /path or http(s):// URL",
                        "example": "https://blog.example.com/"
                    },
                    "code": {
                        "type": "integer",
                        "enum": [
                            301,
                            302,
                            307,
                            308
                        ],
                        "example": 301
                    },
                    "keep_path": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
            "RedirectsResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object",
                        "properties": {
                            "app": {
                                "type": "string",
                                "example": "myapp"
                            },
                            "redirect": {
                                "$ref": "#/components/schemas/AppRedirect"
                            },
                            "redirects": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/RedirectRule"
                                }
                            }
                        }
                    }
                }
            },
            "RedirectSetRequest": {
                "type": "object",
                "required": [
                    "to"
                ],
                "properties": {
                    "to": {
                        "type": "string",
                        "description": "Target URL (http(s)://host[/path])",
                        "example": "https://new.example.com"
                    },
                    "code": {
                        "type": "integer",
                        "enum": [
                            301,
                            302,
                            307,
                            308
                        ],
                        "default": 301
                    },
                    "keep_path": {
                        "type": "boolean",
                        "default": true,
                        "description": "Keep request path and query on the target"
                    }
                }
            },
            "RedirectAddRequest": {
                "type": "object",
                "required": [
                    "from",
                    "to"
                ],
                "properties": {
                    "from": {
                        "type": "string",
                        "description": "Source path, written decoded. Ending in / = prefix match (e.g. /blog/)",
                        "example": "/blog/"
                    },
                    "to": {
                        "type": "string",
                        "description": "Target: same-app /path or http(s):// URL",
                        "example": "https://blog.example.com/"
                    },
                    "code": {
                        "type": "integer",
                        "enum": [
                            301,
                            302,
                            307,
                            308
                        ],
                        "default": 301
                    },
                    "keep_path": {
                        "type": "boolean",
                        "default": true,
                        "description": "For prefix redirects: append the rest of the URI to the target"
                    }
                }
            },
            "RedirectRemoveRequest": {
                "type": "object",
                "required": [
                    "from"
                ],
                "properties": {
                    "from": {
                        "type": "string",
                        "description": "Source path of the redirect to remove",
                        "example": "/blog/"
                    }
                }
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Missing or invalid token",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            }
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Token lacks required ability",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Invalid ability provided."
                                }
                            }
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
...
        }
    },
...
}